// client/src/app/shared/Components/ge-filters/GEFilters.ts
export interface GEFilters {
	days: number;
	from: string;
	to: string;
	dateRange: string;
	periodFrom: string;
	periodTo: string;
	level: string;
	childId: string;
	parentId: string;
	sender: string;
	receiver: string;
	status: string;
	bu: string;
	altId: string;
	notes: string;
	searchBy: string;
	adHoc: string;
	duration: string;
	id: string;
	durationTime: string;
}